Houdini Engine for Unity: updated the GetTime and SetTime methods to use double's instead of float's in order to reflect the same change in HAPI (r511007)

+++ Plugins/HoudiniEngineUnity/Scripts/Sessions/HEU_SessionHAPI.cs
@@ -1253,15 +1253,15 @@ namespace HoudiniEngineUnity
 
         // TIME -----------------------------------------------------------------------------------------------------
 
-        public override float GetTime()
+        public override double GetTime()
         {
-            float time = 0;
+            double time = 0;
             HAPI_Result result = HEU_HAPIFunctions.HAPI_GetTime(ref _sessionData._HAPISession, out time);
             HandleStatusResult(result, "Getting Time", false, true);
             return time;
         }
 
-        public override bool SetTime(float time)
+        public override bool SetTime(double time)
         {
             HAPI_Result result = HEU_HAPIFunctions.HAPI_SetTime(ref _sessionData._HAPISession, time);
             HandleStatusResult(result, "Setting Time", false, true);
